+++ corenlp/src/main/scala/org/clulab/processors/ProcessorShell.scala
@@ -3,9 +3,8 @@ package org.clulab.processors
 import java.io.PrintWriter
 import org.clulab.dynet.Utils
 import org.clulab.processors.clu.CluProcessor
-import org.clulab.processors.clucore.CluCoreProcessor
 import org.clulab.processors.corenlp.CoreNLPProcessor
-import org.clulab.processors.fastnlp.{FastNLPProcessor, FastNLPProcessorWithSemanticRoles}
+import org.clulab.processors.fastnlp.FastNLPProcessorWithSemanticRoles
 import org.clulab.utils.CliReader
 import org.clulab.utils.DefaultMenuItem
 import org.clulab.utils.ExitMenuItem
@@ -23,12 +22,11 @@ object ProcessorShell extends App {
   lazy val core: Processor = new CoreNLPProcessor() // this uses the slower constituent parser
   lazy val fast: Processor = new FastNLPProcessorWithSemanticRoles() // this uses the faster dependency parser
   lazy val clu: Processor = new CluProcessor()
-  lazy val cluCore: Processor = new CluCoreProcessor() // CLU + CoreNLP's numeric entity recognizer
 
   Utils.initializeDyNet()
 
-  var proc = cluCore // The initial proc does not get initialized.
-  val cluCorePrompt = "(clucore)>>> "
+  var proc = clu // The initial proc does not get initialized.
+  val cluCorePrompt = "(clu)>>> "
 
   val lineReader = new CliReader(cluCorePrompt, "user.home", ".processorshellhistory")
   val printWriter = new PrintWriter(System.out)
@@ -51,10 +49,6 @@ object ProcessorShell extends App {
     prepareProcessor("(clu)>>> ", "Preparing CluProcessor...", clu)
   }
 
-  def prepareCluCore(menu: Menu, text: String): Boolean = {
-    prepareProcessor(cluCorePrompt, "Preparing CluCoreProcessor...", cluCore)
-  }
-
   def annotate(menu: Menu, text: String): Boolean = {
     val doc = proc.annotate(text)
     doc.prettyPrint(printWriter)
@@ -67,7 +61,6 @@ object ProcessorShell extends App {
     new MainMenuItem(":core", "use CoreNLPProcessor", prepareCore),
     new MainMenuItem(":fast", "use FastNLPProcessor", prepareFast),
     new MainMenuItem(":clu", "use CluProcessor", prepareClu),
-    new MainMenuItem(":clucore", "use CluCoreProcessor", prepareCluCore),
     new ExitMenuItem(":exit", "exit system")
   )
   val defaultMenuItem = new DefaultMenuItem(annotate)